+++ spec/ruby_spec.rb
@@ -3,6 +3,7 @@ require_relative 'spec_helper'
 describe "Ruby apps" do
   describe "default WEB_CONCURRENCY" do
     it "auto scales WEB_CONCURRENCY" do
+      pending("https://github.com/heroku/api/issues/4426")
       app = Hatchet::Runner.new("default_ruby")
       app.setup!
       app.set_config("SENSIBLE_DEFAULTS" => "enabled")
@@ -27,24 +28,6 @@ describe "Ruby apps" do
       end
     end
 
-    context "Ruby 1.8.7 on cedar" do
-      it "doesn't run rake tasks if no rake gem" do
-        app = Hatchet::Runner.new('mri_187_no_rake').setup!
-        app.heroku.put_stack(app.name, "cedar")
-        app.deploy do |app, heroku|
-          expect(app.output).not_to include("foo")
-        end
-      end
-
-      it "runs a rake task if the gem exists" do
-        app = Hatchet::Runner.new('mri_187_rake').setup!
-        app.heroku.put_stack(app.name, "cedar")
-        app.deploy do |app, heroku|
-          expect(app.output).to include("foo")
-        end
-      end
-    end
-
     context "Ruby 1.9+" do
       it "runs rake tasks if no rake gem" do
         Hatchet::Runner.new('mri_200_no_rake').deploy do |app, heroku|
@@ -55,6 +38,24 @@ describe "Ruby apps" do
       it "runs a rake task if the gem exists" do
         Hatchet::Runner.new('mri_200_rake').deploy do |app, heroku|
           expect(app.output).to include("foo")
+        end
+      end
+    end
+  end
+
+  describe "database configuration" do
+    context "no active record" do
+      it "writes a heroku specific database.yml" do
+        Hatchet::Runner.new("default_ruby").deploy do |app, heroku|
+          expect(app.output).to include("Writing config/database.yml to read from DATABASE_URL")
+        end
+      end
+    end
+
+    context "active record 4.1+" do
+      it "doesn't write a heroku specific database.yml" do
+        Hatchet::Runner.new("activerecord41_scaffold").deploy do |app, heroku|
+          expect(app.output).not_to include("Writing config/database.yml to read from DATABASE_URL")
         end
       end
     end